feat(list-recipes): add styles for list of recipes
