+++ src/recipes/components/index.tsx
@@ -1,0 +1,1 @@
+export { RecipeCard } from "./RecipeCard"
