+++ src/recipes/utils/formatting.ts
@@ -1,0 +1,4 @@
+import { Ingredient } from "../types"
+
+export const formatIngredients = (ingredients: Ingredient[]) =>
+  ingredients.map((ingredient) => ingredient.name).join(", ")
